requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# AI & LLM
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Report Generation (PDF)
markdown>=3.5.0
weasyprint>=60.0

# Streamlit Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# RAG (Retrieval Augmented Generation)
chromadb>=0.4.0
pypdf>=3.17.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# Optional: For future visualization/ML features
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
